implement if condition to show loggin or logout user
