+++ src/containers/Dashboard/Navbar.scss
@@ -13,6 +13,7 @@
       font-family: "sans-serif";
     }
   }
+
   .form-inline .form-control-sm {
     background-color: #252e38 !important;
     color: #ffffff;
@@ -20,7 +21,28 @@
   &.shadow-lg {
     color: #131c25;
   }
-
+  .navbar.navbar-dark.login-btn {
+    padding-top: 3px;
+    padding-bottom: 3px;
+    width: 145px;
+    .btn-login {
+      border-radius: 0%;
+      padding-bottom: 10px;
+      padding-top: 10px;
+      padding-left: 40px;
+      padding-right: 40px;
+      background-color: hsla(0, 0%, 100%, 0.08);
+      font-weight: 550;
+      text-transform: uppercase;
+      border: black;
+      height: 42px;
+      font-family: "sans-serif";
+      color: white;
+      &:hover {
+        text-decoration: none;
+      }
+    }
+  }
   .nav-item {
     input {
       border-color: transparent;
