+++ src/routes/index.js
@@ -26,29 +26,40 @@ import {
   PROFILE
 } from "../constants";
 import { BrowserRouter as Router } from "react-router-dom";
-import {getItem} from '../services/Session'
-// import Header from '../components/Header'
+import { getItem } from "../services/Session";
 
 class Routes extends Component {
+  constructor(props){
+    super(props);
+    this.state={
+      refresh:false
+    }
+    this.onRefresh = this.onRefresh.bind(this);
+  }
+
+  onRefresh() {
+    this.setState({
+      refresh: false
+    });
+  }
 
   render() {
-    // debugger;
+    debugger;
     const name = getItem("name");
     return (
       <Router>
-           {/* <Header btnName="Log In" redirectTo={LOGIN} /> */}
-          <Navbar name={name}/>
-          <GuestRoute exact path={HOME} component={Home} />
-          <GuestRoute path={SINGUP} component={SignUp} />
-          <GuestRoute path={LOGIN} component={Login} />
-          <ProtectedRoute path={DASHBOARD} component={Dashboard} />
-          <ProtectedRoute path={BOOKMARK} component={Bookmark} />
-          <ProtectedRoute path={INTERVIEW_PREP_KIT} component={InterviewPrep} />
-          <ProtectedRoute path={PROBLEM_SOLVING} component={ProblemSolving} />
-          <ProtectedRoute path={STATISTICS} component={Statistics} />
-          <ProtectedRoute path={JAVASCRIPT} component={Javascript} />
-          <ProtectedRoute path={CODE} component={Code} />
-          <ProtectedRoute path={PROFILE} component={Profile} />
+        <Navbar onRefresh={this.onRefreshs} name={name} />
+        <GuestRoute exact path={HOME} component={Home} />
+        <GuestRoute path={SINGUP} component={SignUp} />
+        <GuestRoute path={LOGIN} component={Login} />
+        <ProtectedRoute path={DASHBOARD} component={Dashboard} />
+        <ProtectedRoute path={BOOKMARK} component={Bookmark} />
+        <ProtectedRoute path={INTERVIEW_PREP_KIT} component={InterviewPrep} />
+        <ProtectedRoute path={PROBLEM_SOLVING} component={ProblemSolving} />
+        <ProtectedRoute path={STATISTICS} component={Statistics} />
+        <ProtectedRoute path={JAVASCRIPT} component={Javascript} />
+        <ProtectedRoute path={CODE} component={Code} />
+        <ProtectedRoute path={PROFILE} component={Profile} />
       </Router>
     );
   }
